Add base smart sense board config
Added in a custom board config for smart sense based off the P2771 board
fom NVidia.

*Ported from Smart Sense Repo*

u-boot configures LM96063 (LM63) for tach

The LM96063 power on default is the tach function is disabled. The
current linux kernel driver has no facility for enabling or disabling
this at all; it only checks to see if the tach function is enabled. In
fact, comments in the driver code say it's expecting the BIOS to have
set up the LM63-compatible device.

Enable the tach function in the chip. This causes the driver to add the
fan1_* device entries, which allows monitoring of the tach.

NOTE: i2c bus numbering is differeing in u-boot than in the linux
kernel. I do not know why.

Use slot GPIO and process DTB

Add code to read the slot GPIO line to determine the slot value

Make the tachometer initialisation happen only on Slot A, which is where
the tach is connected. Slot B no longer pops an error during boot.

Add the code in ft_board_setup() to process the DTB for Smartsense slot
designations. Designed extensible for future products with even moar
Jetsons.